Refactor width, height, and center to be double types.

+++ shape.cpp
@@ -10,22 +10,22 @@ Shape::Shape():
     _center(306, 396), _height(72), _width(72)
 {}
 
-pair<int, int> Shape::getCenter()
+pair<double, double> Shape::getCenter()
 {
     return _center;
 }
 
-int Shape::getHeight()
+double Shape::getHeight()
 {
     return _height;
 }
 
-int Shape::getWidth()
+double Shape::getWidth()
 {
     return _width;
 }
 
-void Shape::setCenter(int x, int y)
+void Shape::setCenter(double x, double y)
 {
     _center=std::make_pair(x, y);
 }
@@ -38,7 +38,7 @@ Circle::Circle():
     _radius(getHeight() / 2)
 {}
 
-int Circle::getRadius()
+double Circle::getRadius()
 {
     return _radius;
 }
